Move entities under database/entities
